Moved road to its own file and added a verifier that ensures the vehicle number inputs match the lane setup.
